created printing components for document tabs
