+++ website/src/pages/documents/print-layout.tsx
@@ -1,55 +1,12 @@
 import React, { useEffect, useState } from "react"
 import { Helmet } from "react-helmet"
 import { Breadcrumbs, Link } from "src/components"
-import { FormProvider as FormProviderParagraph } from "src/edit-paragraph-form-context"
-import { EditWordCheckProvider } from "src/edit-word-check-context"
-import { FormProvider } from "src/edit-word-form-context"
 import * as Dailp from "src/graphql/dailp"
-import { usePreferences } from "src/preferences-context"
 import { useLocation } from "src/renderer/PageShell"
 import { fonts } from "src/style/constants"
-import { LevelOfDetail } from "src/types"
-import { DocumentContents } from "./document.page"
 import * as css from "./print-layout.css"
 
 export const printDocument = () => window.print()
-
-export const PrintLegend = ({
-  levelOfDetail,
-}: {
-  levelOfDetail: LevelOfDetail
-}) => (
-  <div className={css.printLegendBox}>
-    <h2 className={css.printHeading}>How To Read This Document</h2>
-    <p className={css.printLegendSyllabary}>
-      <strong>Syllabary Line:</strong> Text from the original source
-    </p>
-    {levelOfDetail >= LevelOfDetail.Pronunciation && (
-      <p className={css.printLegendItem}>
-        <strong>Phonetics Line:</strong> Pronunciation of a word
-      </p>
-    )}
-    {levelOfDetail >= LevelOfDetail.Segmentation && (
-      <p className={css.printLegendItem}>
-        <strong>Word Parts Line:</strong> Separates out meaningful parts of a
-        word
-      </p>
-    )}
-    {levelOfDetail >= LevelOfDetail.Segmentation && (
-      <p className={css.printLegendItem}>
-        <strong>Gloss Line:</strong> Provides the meaning of separated word
-        parts
-      </p>
-    )}
-    <p className={css.printLegendItem}>
-      <strong>Translation Line:</strong> An English translation for a word
-    </p>
-    <p className={css.printLegendItem}>
-      <strong>Paragraph Translation:</strong> Provides a free translation of a
-      section of text
-    </p>
-  </div>
-)
 
 export const usePrintFooterContent = () => {
   const location = useLocation()
@@ -62,7 +19,7 @@ export const usePrintFooterContent = () => {
   return { date, url }
 }
 
-export const PrintDocument = (p: {
+export const PrintLayout = (p: {
   doc: Dailp.DocumentFieldsFragment
   breadcrumbs?: readonly Pick<
     Dailp.CollectionChapter["breadcrumbs"][0],
@@ -70,8 +27,8 @@ export const PrintDocument = (p: {
   >[]
   rootPath?: string
   collectionTitle?: string
+  children?: React.ReactNode
 }) => {
-  const { levelOfDetail, cherokeeRepresentation } = usePreferences()
   const slug = p.doc.slug ?? ""
   const [{ data }] = Dailp.useDocumentDetailsQuery({ variables: { slug } })
   const contributorNames =
@@ -148,36 +105,8 @@ export const PrintDocument = (p: {
           {p.doc.title}
           {p.doc.date && ` (${p.doc.date.year})`}
         </h1>
-        <PrintLegend levelOfDetail={levelOfDetail} />
       </div>
-      <h2 className={css.printSectionHeading}>
-        Translation
-        <span className={css.printSectionHeadingRule} />
-      </h2>
-      <div
-        className={
-          levelOfDetail >= LevelOfDetail.Segmentation
-            ? `${css.printBodyContent} ${css.printHideParagraphTranslation}`
-            : css.printBodyContent
-        }
-      >
-        <EditWordCheckProvider>
-          <FormProvider>
-            <FormProviderParagraph>
-              <DocumentContents
-                doc={p.doc}
-                levelOfDetail={levelOfDetail}
-                cherokeeRepresentation={cherokeeRepresentation}
-                openDetails={() => {}}
-                wordPanelDetails={{
-                  currContents: null,
-                  setCurrContents: () => {},
-                }}
-              />
-            </FormProviderParagraph>
-          </FormProvider>
-        </EditWordCheckProvider>
-      </div>
+      {p.children}
       <div className={css.printBlankPage}>
         This page was left intentionally blank.
       </div>
